# models.py
# generated by datamodel-codegen:
#   filename:  openapi.yaml
#   timestamp: 2025-04-17T23:17:17+00:00

from __future__ import annotations

from typing import Optional

from pydantic import AnyUrl, BaseModel, EmailStr


class UserRegisterRequest(BaseModel):
    username: str
    email: EmailStr
    password: str


class UserLoginRequest(BaseModel):
    username: str
    password: str


class UserResponse(BaseModel):
    user_id: str
    username: str
    email: str


class LoginResponse(BaseModel):
    access_token: str
    token_type: str


class ProjectCreateRequest(BaseModel):
    name: str
    description: Optional[str] = None


class ProjectResponse(BaseModel):
    project_id: str
    name: str
    description: Optional[str] = None


class ProjectListResponse(BaseModel):
    projects: list[ProjectResponse]


class TaskStatusResponse(BaseModel):
    task_id: str
    status: str
    result: Optional[str] = None


class WebhookRegisterRequest(BaseModel):
    url: AnyUrl
    event: str


class WebhookResponse(BaseModel):
    webhook_id: str
    url: AnyUrl
    event: str


class WebsiteAnalyticsResponse(BaseModel):
    website_id: str
    visits: int
    bounce_rate: float
    avg_time_on_site: float


class WebsiteMonetizationResponse(BaseModel):
    website_id: str
    revenue: float
    currency: str
    sources: list[str]